Update input type to int for range check
Changed input type from float to int for better validation.

+++ 20.py
@@ -1,5 +1,5 @@
 # 20. Leia um valor e: verifique se eles está entre 0 e 100, caso o número esteja fora do intervalo, mostre na tela o valor.
-n = float(input())
+n = int(input())
 if n >= 0 and n <= 100:
     print("Dentro do intervalo")
 else:
